requirements.txt
diskcache==5.6.3
numpy==2.1.2
pandas==2.2.3
protobuf==3.20.3
Requests==2.32.3
tenacity==8.2.2
tqdm==4.66.4
python-dotenv==0.21.0
google-cloud-bigquery
google-auth
pyarrow
gunicorn